+++ tree_hugger/core/__init__.py
@@ -1,3 +1,3 @@
 from .code_parser import BaseParser
-from .python_parser import PythonParser
+from .parser.python.python_parser import PythonParser
 from .code_parser import match_from_span
